Create Flight Management
Talking about the features of this system, it contains only the admin section. This project is in Python 3.8.1. This is a management system that is made in order to keep a track record of flights, staff, and updates on flight details. This project mainly focuses on CRUD with search functionality. The design is so simple that the user won’t find any difficulties while working on it.